Refactor Command defines into ClimateChamberWrapperDefines.h
